+++ src/frontend/hooks/useBom/boid-query-map.js
@@ -1,0 +1,10 @@
+import {
+  // GET_APM_ORGANIZATIONS,
+  GET_PATHWAYS_ORGANIZATIONS,
+  // GET_PAYER_ORGANIZATIONS,
+  // GET_PROVIDER_ORGANIZATIONS,
+} from './../../api/queries'
+
+export default {
+  '5e46a9784a5ab07ddbe3dc88': GET_PATHWAYS_ORGANIZATIONS
+}